requirements.txt
openai==1.76.0
python-dotenv==1.0.1
python-docx==1.1.2
pypdf==4.3.1
rank-bm25==0.2.2
numpy==1.26.4
jinja2==3.1.6
typer==0.12.3
rich==13.7.1
pydantic==2.8.2
python-pptx==1.0.2
